Refactor UpdateFields and inline ItemUpdater class into it
