+++ lib/dynamoid/transactions/mutation.rb
@@ -9,7 +9,6 @@ require_relative 'mutation/update_fields'
 require_relative 'mutation/update_attributes'
 require_relative 'mutation/upsert'
 require_relative 'mutation/inc'
-require_relative 'mutation/item_updater'
 require_relative 'mutation/touch'
 require_relative 'mutation/increment'
 require_relative 'mutation/import'
